Add launch files + formatting
