+++ moveit_ros/moveit_servo/config/demo_rviz_config.rviz
