+++ moveit_ros/moveit_servo/demos/cpp_interface/demo_joint_jog.cpp
@@ -69,7 +69,8 @@ int main(int argc, char* argv[])
   auto servo = Servo(servo_node, servo_param_listener);
 
   // Wait for some time, so that we can actually see when the robot moves.
-  std::this_thread::sleep_for(std::chrono::seconds(5));
+  // This is just for convenience, should not be used for sync in real application.
+  std::this_thread::sleep_for(std::chrono::seconds(3));
 
   // Frquency at which the commands will be send to robot controller.
   rclcpp::WallRate rate(1.0 / servo_params.publish_period);
