+++ moveit_ros/moveit_servo/demos/cpp_interface/demo_twist.cpp
@@ -68,6 +68,10 @@ int main(int argc, char* argv[])
   // Create the servo object
   auto servo = Servo(servo_node, servo_param_listener);
 
+  // Wait for some time, so that we can actually see when the robot moves.
+  // This is just for convenience, should not be used for sync in real application.
+  std::this_thread::sleep_for(std::chrono::seconds(3));
+
   // Frquency at which the commands will be send to robot controller.
   rclcpp::WallRate rate(1.0 / servo_params.publish_period);
