+++ moveit_ros/moveit_servo/launch/demo_joint_jog.launch.py
@@ -2,18 +2,11 @@ import os
 import launch
 import launch_ros
 from ament_index_python.packages import get_package_share_directory
-from launch.conditions import IfCondition, UnlessCondition
-from launch.substitutions import LaunchConfiguration
 from launch_param_builder import ParameterBuilder
 from moveit_configs_utils import MoveItConfigsBuilder
 
 
 def generate_launch_description():
-    # Launch Servo as a standalone node or as a "node component" for better latency/efficiency
-    launch_as_standalone_node = LaunchConfiguration(
-        "launch_as_standalone_node", default="false"
-    )
-
     moveit_config = (
         MoveItConfigsBuilder("moveit_resources_panda")
         .robot_description(file_path="config/panda.urdf.xacro")
@@ -84,19 +77,6 @@ def generate_launch_description():
         package="rclcpp_components",
         executable="component_container_mt",
         composable_node_descriptions=[
-            # Example of launching Servo as a node component
-            # Launching as a node component makes ROS 2 intraprocess communication more efficient.
-            launch_ros.descriptions.ComposableNode(
-                package="moveit_servo",
-                plugin="moveit_servo::ServoNode",
-                name="servo_node",
-                parameters=[
-                    servo_params,
-                    moveit_config.robot_description,
-                    moveit_config.robot_description_semantic,
-                ],
-                condition=UnlessCondition(launch_as_standalone_node),
-            ),
             launch_ros.descriptions.ComposableNode(
                 package="robot_state_publisher",
                 plugin="robot_state_publisher::RobotStatePublisher",
@@ -109,25 +89,14 @@ def generate_launch_description():
                 name="static_tf2_broadcaster",
                 parameters=[{"child_frame_id": "/panda_link0", "frame_id": "/world"}],
             ),
-            launch_ros.descriptions.ComposableNode(
-                package="moveit_servo",
-                plugin="moveit_servo::JoyToServoPub",
-                name="controller_to_servo_node",
-            ),
-            launch_ros.descriptions.ComposableNode(
-                package="joy",
-                plugin="joy::Joy",
-                name="joy_node",
-            ),
         ],
         output="screen",
     )
-
-    # Optionally launch a standalone Servo node.
+    # Launch a standalone Servo node.
     # As opposed to a node component, this may be necessary (for example) if Servo is running on a different PC
     servo_node = launch_ros.actions.Node(
         package="moveit_servo",
-        executable="servo_node_main",
+        executable="demo_joint_jog",
         parameters=[
             servo_params,
             low_pass_filter_coeff,
@@ -136,7 +105,6 @@ def generate_launch_description():
             moveit_config.robot_description_kinematics,
         ],
         output="screen",
-        condition=IfCondition(launch_as_standalone_node),
     )
 
     return launch.LaunchDescription(
